move normalize to the more general known_instances.js
